More NSUInteger fixes.
git-svn-id: http://svn.gna.org/svn/etoile/trunk/Etoile@7983 7b3f36aa-e6db-49c1-8b4c-66eac1997e64

+++ Services/Private/AZSwitch/AZSwitchingWindow.m
@@ -61,7 +61,7 @@
 }
 
 - (id) initWithContentRect: (NSRect)contentRect
-                 styleMask: (unsigned int)aStyle
+                 styleMask: (NSUInteger)aStyle
                    backing: (NSBackingStoreType)bufferingType
                      defer: (BOOL)flag
                     screen: (NSScreen*)aScreen
