+++ Services/User/Typewriter/TWCharacterPanel.m
@@ -105,7 +105,7 @@ static TWCharacterPanel *sharedInstance;
 }
 
 - (id) initWithContentRect: (NSRect)contentRect
-                 styleMask: (unsigned int)aStyle
+                 styleMask: (NSUInteger)aStyle
 	           backing: (NSBackingStoreType)bufferingType
 	             defer: (BOOL)flag
 	            screen: (NSScreen*)aScreen
